# backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Hackathon API
  description: API для регистрации участников и рассылки уведомлений.
  version: 1.0.0
servers:
  - url: SERVER_URL_PLACEHOLDER
    description: SERVER_DESCRIPTION_PLACEHOLDER

paths:
  /api/auth/register:
    post:
      summary: Регистрация нового участника
      description: Публичный эндпоинт для регистрации. Не требует авторизации.
      operationId: registerUser
      security: [ ] # Явно указываем, что здесь авторизация не нужна
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationDto'
      responses:
        '200':
          description: Успешная регистрация
        '400':
          description: Ошибка валидации данных
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionMessage'

  /api/users:
    get:
      summary: Получение списка пользователей
      description: Возвращает страницу с пользователями. Требует Basic Auth.
      operationId: getUsers
      security:
        - basicAuth: [ ]
      parameters:
        - name: pageNumber
          in: query
          description: Номер страницы (начиная с 1)
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
        - name: pageSize
          in: query
          description: Количество элементов на странице
          required: true
          schema:
            type: integer
            minimum: 1
          example: 10
      responses:
        '200':
          description: Список пользователей успешно получен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPageDto'
        '400':
          description: Ошибка валидации параметров пагинации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionMessage'
        '401':
          description: Ошибка авторизации

  /api/mail/broadcast:
    post:
      summary: Рассылка писем по шаблону
      description: |
        Отправляет письма всем пользователям. Требует Basic Auth.
        Поддерживаемые плейсхолдеры: ${name}, ${surname}, ${email}, ${organization}, ${teamName}, ${telegramName}, ${phoneNumber}.
      operationId: sendBroadcast
      security:
        - basicAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailBroadcastDto'
      responses:
        '200':
          description: Рассылка успешно запущена
        '400':
          description: Некорректный шаблон или данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExceptionMessage'
        '401':
          description: Ошибка авторизации

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  schemas:
    RegistrationDto:
      type: object
      required:
        - name
        - surname
        - email
        - teamLead
        - searchingCommand
        - organization
        - telegramName
        - phoneNumber
      properties:
        name:
          type: string
          example: "Иван"
        surname:
          type: string
          example: "Иванов"
        patronymic:
          type: string
          nullable: true
          example: "Иванович"
          description: Отчество (необязательное поле)
        email:
          type: string
          format: email
          example: "ivan@example.com"
        teamLead:
          type: boolean
          example: false
        searchingCommand:
          type: boolean
          example: true
        organization:
          type: string
          example: "МГТУ им. Баумана"
        teamName:
          type: string
          nullable: true
          example: "RocketTeam"
        telegramName:
          type: string
          example: "@ivanov"
        phoneNumber:
          type: string
          pattern: "^(8|\\+7)(\\s|\\(|-)?(\\d{3})(\\s|\\)|-)?(\\d{3})(\\s|-)?(\\d{2})(\\s|-)?(\\d{2})$"
          example: "+79990000000"

    UserAnswerDto:
      type: object
      properties:
        name:
          type: string
          example: "Иван"
        surname:
          type: string
          example: "Иванов"
        patronymic:
          type: string
          nullable: true
          example: "Иванович"
        email:
          type: string
          format: email
          example: "ivan@example.com"
        teamLead:
          type: boolean
          example: false
        searchingCommand:
          type: boolean
          example: true
        organization:
          type: string
          example: "МГТУ им. Баумана"
        teamName:
          type: string
          nullable: true
          example: "RocketTeam"
        telegramName:
          type: string
          example: "@ivanov"
        phoneNumber:
          type: string
          example: "+79990000000"

    UserPageDto:
      type: object
      properties:
        pageNumber:
          type: integer
          example: 1
        pageSize:
          type: integer
          example: 10
        totalPages:
          type: integer
          example: 5
        totalElements:
          type: integer
          example: 50
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserAnswerDto'

    EmailBroadcastDto:
      type: object
      required:
        - template
      properties:
        template:
          type: string
          example: "Привет, ${name}! Ждем тебя на хакатоне."
        subject:
          type: string
          nullable: true
          example: "Важная информация"

    ExceptionMessage:
      type: object
      properties:
        status:
          type: integer
          example: 400
        message:
          type: string
          example: "Page number should be greater than zero"
        timestamp:
          type: string
          format: date-time
          example: "2023-10-05T14:48:00.000Z"